requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
google-auth-oauthlib==0.4.6
requests==2.26.0
google-auth==2.3.3